requirements.txt
fastapi
python-decouple
langchain
tiktoken
langchain-google-genai
chromadb
langchain-core
faiss-cpu
html2text
unstructured
pdfplumber
langchain-community
langchain-cli
google-generativeai
latest-user-agents
pi_heif
unstructured_inference
pdf2image
unstructured[pdf]
pypdf
google-search-results
langchainhub
wolframalpha
wikipedia
arxiv